Move test files, cleanup imports
