+++ packages/shared/src/utils/get-collection-type.test.ts
@@ -1,5 +1,6 @@
-import { getCollectionType } from '../../src/utils';
-import { Collection } from '../../src/types';
+import { describe, expect, it } from 'vitest';
+import { Collection } from '../types';
+import { getCollectionType } from './get-collection-type';
 
 const TableCollection: Collection = {
 	collection: 'table',
