+++ packages/shared/src/utils/get-relation-type.test.ts
@@ -1,5 +1,6 @@
-import { getRelationType } from '../../src/utils/get-relation-type';
-import { Relation } from '../../src/types';
+import { describe, expect, it } from 'vitest';
+import { Relation } from '../types';
+import { getRelationType } from './get-relation-type';
 
 describe('getRelationType', () => {
 	it('returns m2o when relation is the same as collection and field', () => {
