+++ packages/shared/src/utils/move-in-array.test.ts
@@ -1,4 +1,5 @@
-import { moveInArray } from '../../src/utils/move-in-array';
+import { describe, expect, it } from 'vitest';
+import { moveInArray } from './move-in-array';
 
 describe('moveInArray', () => {
 	const testArray = [1, 2, 3, 4, 5, 6];
