+++ packages/shared/src/utils/pluralize.test.ts
@@ -1,4 +1,5 @@
-import { pluralize, depluralize } from '../../src/utils/pluralize';
+import { describe, expect, it } from 'vitest';
+import { depluralize, pluralize } from './pluralize';
 
 describe('pluralize', () => {
 	it('adds an s to the end of the string', () => {
