+++ packages/shared/src/utils/validate-extension-manifest.test.ts
@@ -1,4 +1,5 @@
-import { validateExtensionManifest } from '../../src/utils/validate-extension-manifest';
+import { describe, expect, it } from 'vitest';
+import { validateExtensionManifest } from './validate-extension-manifest';
 
 describe('', () => {
 	it('returns false when passed item has no name or version', () => {
